Set all SmallUtilities individual packs to GitHub I've also turned around the conflicts statements, so each individual pack conflicts with the complete pack, making it easier to check for missing conflicts.
I have set the ksp_version for each mod to the current release situation. As packs are moved to 1.1, all you need do is change the ksp_version.

If in the future, you wish to set a range of ksp versions, you can replace the individual "ksp_version" line with a min/max pair, like this:

    "ksp_version_min": "1.1.0",
    "ksp_version_max": "1.1.5",